# openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Future Me API",
    "description": "Future Me is a time capsule letter platform that allows users to send letters to their future self or loved ones, scheduled for delivery on a future date. This API provides endpoints for user authentication, letter creation and management, draft system, scheduled delivery, notifications, and media attachments.",
    "version": "1.0.0",
    "contact": {
      "name": "Future Me Support",
      "email": "support@futureme.com"
    }
  },
  "servers": [
    {
      "url": "https://backend.hyfuture.app/api/v1",
      "description": "Production Server"
    },
    {
      "url": "http://localhost:4001/api/v1",
      "description": "Local Development Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT access token. Get it from the login or verify-auth-code endpoints."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        }
      },
...
      "Attachment": {
        "type": "object",
        "required": ["fileUrl", "type"],
        "properties": {
          "fileUrl": {
            "type": "string",
            "format": "uri",
            "example": "https://example.com/files/photo.jpg",
            "description": "URL to the attached file"
          },
          "type": {
            "type": "string",
            "enum": ["IMAGE", "VIDEO", "AUDIO", "DOCUMENT"],
            "example": "IMAGE",
            "description": "Type of attachment"
          }
        }
      },
      "CreateLetterRequest": {
        "type": "object",
        "required": ["recipientEmail", "deliveryDate"],
        "properties": {
          "subject": {
            "type": "string",
            "maxLength": 200,
            "example": "A Letter to My Future Self",
            "description": "Optional subject line for the letter"
          },
          "content": {
            "type": "string",
            "maxLength": 10000,
            "example": "Dear Future Me, I hope you're doing well...",
            "description": "Optional letter content (supports rich text)"
          },
          "recipientEmail": {
            "type": "string",
            "format": "email",
            "example": "future@example.com",
            "description": "Email address of the recipient"
          },
          "recipientName": {
            "type": "string",
            "example": "Future Me",
            "description": "Optional name of the recipient"
          },
          "deliveryDate": {
            "type": "string",
            "format": "date-time",
            "example": "2026-12-25T10:00:00Z",
            "description": "ISO 8601 date-time when the letter should be delivered"
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attachment"
            },
            "description": "Optional media attachments (photos, videos, audio, documents)"
          },
          "isPublic": {
            "type": "boolean",
            "example": true,
            "description": "If true, the letter will be visible on the public letters feed."
          }
        }
      },
      "CreateGuestLetterRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CreateLetterRequest"
          },
          {
            "type": "object",
            "required": ["senderEmail"],
            "properties": {
              "senderEmail": {
                "type": "string",
                "format": "email",
                "example": "guest@example.com",
                "description": "Email address of the guest sender (required for guest letters)"
              },
              "senderName": {
                "type": "string",
                "example": "John Doe",
                "description": "Optional name of the sender"
              }
            }
          }
        ]
      },
      "UpdateLetterRequest": {
        "type": "object",
        "properties": {
          "subject": {
            "type": "string",
            "maxLength": 200,
            "example": "Updated Subject"
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000,
            "example": "Updated content..."
          },
          "recipientEmail": {
            "type": "string",
            "format": "email",
            "example": "newrecipient@example.com"
          },
          "recipientName": {
            "type": "string",
            "example": "New Recipient Name"
          },
          "deliveryDate": {
            "type": "string",
            "format": "date-time",
            "example": "2027-01-01T10:00:00Z"
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attachment"
            },
            "description": "Replace all attachments. Provide empty array to remove all attachments."
          },
          "isPublic": {
            "type": "boolean",
            "example": true,
            "description": "If true, the letter will be visible on the public letters feed."
          }
        }
      },
      "Letter": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "subject": {
            "type": "string",
            "nullable": true,
            "example": "A Letter to My Future Self"
          },
          "content": {
            "type": "string",
            "example": "Dear Future Me..."
          },
          "senderEmail": {
            "type": "string",
            "format": "email",
            "example": "sender@example.com"
          },
          "senderName": {
            "type": "string",
            "nullable": true,
            "example": "John Doe"
          },
          "recipientEmail": {
            "type": "string",
            "format": "email",
            "example": "recipient@example.com"
          },
          "recipientName": {
            "type": "string",
            "nullable": true,
            "example": "Future Me"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "example": "550e8400-e29b-41d4-a716-446655440001"
          },
          "deliveryDate": {
            "type": "string",
            "format": "date-time",
            "example": "2026-12-25T10:00:00Z"
          },
          "deliveredAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2026-12-25T10:00:00Z"
          },
          "status": {
            "type": "string",
            "enum": ["DRAFT", "SCHEDULED", "DELIVERED", "FAILED"],
            "example": "SCHEDULED"
          },
          "isGuest": {
            "type": "boolean",
            "example": false
          },
          "isPublic": {
            "type": "boolean",
            "example": true,
            "description": "True if this letter is visible on the public letters feed."
          },
          "locked": {
            "type": "boolean",
            "example": true,
            "description": "True if letter is locked (delivery date not reached). Only basic info returned when locked."
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attachment"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T10:00:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T10:00:00Z"
          }
        }
      },
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          },
          "data": {
            "type": "object",
            "description": "Response data (varies by endpoint)"
          },
          "statusCode": {
            "type": "integer",
            "example": 200
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/letters/guest": {
      "post": {
        "tags": ["Letters"],
        "summary": "Create Guest Letter",
        "description": "Create a letter as a guest user. Guests can send only 1 letter per recipient email. Rate limited to 10 requests per hour. Guest letters are automatically linked to user accounts when the guest signs up with the same email.",
        "operationId": "createGuestLetter",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGuestLetterRequest"
              },
              "example": {
                "senderEmail": "guest@example.com",
                "senderName": "John Doe",
                "subject": "A Letter to My Future Self",
                "content": "Dear Future Me, I hope you're doing well...",
                "recipientEmail": "future@example.com",
                "recipientName": "Future Me",
                "deliveryDate": "2026-12-25T10:00:00Z",
                "attachments": [
                  {
                    "fileUrl": "https://example.com/files/photo.jpg",
                    "type": "IMAGE"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Guest letter created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "success": true,
                  "message": "Guest letter created successfully",
                  "data": {
                    "id": "550e8400-e29b-41d4-a716-446655440000",
                    "subject": "A Letter to My Future Self",
                    "content": "Dear Future Me...",
                    "status": "SCHEDULED",
                    "deliveryDate": "2026-12-25T10:00:00Z"
                  },
                  "statusCode": 201
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Guest has already sent a letter to this recipient. Sign up to send more.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 409,
                  "message": "You have already sent a letter to this recipient. Please sign up to send more letters.",
                  "error": "Conflict"
                }
              }
            }
          }
        }
      }
    },
    "/letters": {
      "post": {
        "tags": ["Letters"],
        "summary": "Create Authenticated Letter",
        "description": "Create a letter as an authenticated user. Rate limited to 20 requests per hour. Letter is scheduled for delivery on the specified date.",
        "operationId": "createAuthenticatedLetter",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLetterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Letter created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Letters"],
        "summary": "Get All Letters",
        "description": "Retrieve all letters for the authenticated user. Letters with delivery dates in the future are returned in locked format (minimal info). Optionally filter by status (DRAFT, SCHEDULED, DELIVERED, FAILED).",
        "operationId": "getAllLetters",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["DRAFT", "SCHEDULED", "DELIVERED", "FAILED"]
            },
            "description": "Filter letters by status",
            "example": "SCHEDULED"
          }
        ],
        "responses": {
          "200": {
            "description": "Letters retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "success": true,
                  "message": "Letters retrieved successfully",
                  "data": [
                    {
                      "id": "550e8400-e29b-41d4-a716-446655440000",
                      "deliveryDate": "2026-12-25T10:00:00Z",
                      "status": "SCHEDULED",
                      "locked": true,
                      "createdAt": "2024-01-01T10:00:00Z",
                      "updatedAt": "2024-01-01T10:00:00Z"
                    },
                    {
                      "id": "550e8400-e29b-41d4-a716-446655440001",
                      "subject": "Past Letter",
                      "content": "Full content...",
                      "status": "DELIVERED",
                      "locked": false,
                      "attachments": []
                    }
                  ],
                  "statusCode": 200
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/letters/draft": {
      "post": {
        "tags": ["Letters"],
        "summary": "Create Draft Letter",
        "description": "Create a letter as a draft. Drafts can be updated later before sending. Drafts are always visible to the user (not locked).",
        "operationId": "createDraft",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLetterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Draft saved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/letters/{id}": {
      "get": {
        "tags": ["Letters"],
        "summary": "Get Letter by ID",
        "description": "Retrieve a specific letter by ID. Only accessible if the delivery date has been reached. Returns 403 if delivery date hasn't been reached yet.",
        "operationId": "getLetterById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Letter ID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "Letter retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Delivery date has not been reached yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 403,
                  "message": "You cannot view this letter because the delivery date has not been reached yet",
                  "error": "Forbidden"
                }
              }
            }
          },
          "404": {
            "description": "Letter not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Letters"],
        "summary": "Update Draft Letter",
        "description": "Update a draft letter. Only letters with DRAFT status can be updated. Sent letters cannot be modified. Attachments are replaced if provided.",
        "operationId": "updateDraft",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Letter ID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLetterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Draft updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Cannot update sent letter or invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Not authorized to update this letter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Letter not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Letters"],
        "summary": "Delete Letter",
        "description": "Delete a letter. Only letters with delivery dates that have passed can be deleted. Returns 400 if delivery date hasn't passed yet.",
        "operationId": "deleteLetter",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Letter ID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "Letter deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Cannot delete letter before delivery date",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 400,
                  "message": "You cannot delete this letter because the delivery date has not passed yet",
                  "error": "Bad Request"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Not authorized to delete this letter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Letter not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/letters/public/{id}": {
      "get": {
        "tags": ["Letters"],
        "summary": "Get a public letter by ID",
        "description": "Retrieve a single public letter by its ID. This endpoint is not authenticated and only returns letters that have been marked as public.",
        "operationId": "getPublicLetterById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the public letter"
          }
        ],
        "responses": {
          "200": {
            "description": "Public letter found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Letter"
                }
              }
            }
          },
          "404": {
            "description": "Letter not found or not public",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}